requirements.txt
# Core GUI Framework
PySide6>=6.5.0

# Screen Capture
mss>=9.0.1

# OCR Processing
easyocr>=1.7.0
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# Web Automation (for direct image search)
selenium>=4.10.0
webdriver-manager>=4.0.0

# Clipboard Operations
pyperclip>=1.8.2

# Global Hotkeys
pynput>=1.7.6

# Optional: Windows clipboard support (if pyperclip fails)
# pywin32>=306; sys_platform == 'win32'